rename to lr-server.js to make more sense
